pyproject.toml
[build-system]
requires = ["setuptools >= 61.2.0"]
build-backend = "setuptools.build_meta"

[project]
name = "xcube_stac"
dynamic = ["version"]
authors = [
  {name = "Konstantin Ntokas", email = "konstantin.ntokas@brockmann-consult.de"}
]
description = """\
  xcube data store for accessing stac catalogs\
  """
keywords = [
  "analysis ready data", "data science", "datacube", "xarray", "zarr", "xcube", "stac"
]
readme = {file = "README.md", content-type = "text/markdown"}
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "boto3==1.37.3",
    "dask",
    "h5netcdf",
    "libgdal-jp2openjpeg",
    "numpy",
    "pandas",
    "pyproj",
    "pystac",
    "pystac-client",
    "rasterio",
    "requests",
    "rioxarray",
    "xarray",
    "xmltodict",
    "xcube-core>=1.11.0"
]

[tool.setuptools.dynamic]
version = {attr = "xcube_stac.__version__"}

[tool.setuptools.packages.find]
exclude = [
  "test*",
  "doc*"
]

[project.optional-dependencies]
dev = [
  "black",
  "flake8",
  "pytest",
  "pytest-cov",
  "pytest-recording"
]

[project.urls]
Repository = "https://github.com/xcube-dev/xcube-stac"
Issues = "https://github.com/xcube-dev/xcube-stac/issues"
Changelog = "https://github.com/xcube-dev/xcube-stac/blob/main/CHANGES.md"
